//nar/narinfo: move into here
There's really no concept of libstore in go-nix. As .narinfo files
explicitly describe metadata of .nar files, nar/narinfo seems a much
more intuitive location.
